requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file dev.txt dev.in
#
bcrypt==3.1.2
cffi==1.9.1
django-environ==0.4.1
django-extensions==1.7.5
django-registration-redux==1.4
django==1.10.5
djangorestframework==3.5.3
olefile==0.44
pillow==4.0.0
pycparser==2.17
requests==2.13.0
six==1.10.0
